Isolate v1 project context loader (#2120)
